requirements.txt
gensim==4.0.0
Keras==2.4.3
numpy==1.20.3
Pillow==8.2.0
scipy==1.6.0
sklearn==0.0
torch==1.7.1
torchvision==0.8.2
